Meet C++11 (#335)
* Replaced `__max` to `std::max` and `__min` to `std::min`. Removed project macros `__max` and `__min`.

* Replaced `__max3` to `std::max` and `__min3` to `std::min` or `std::min_element`. Removed project-wide macros `__min3` and `__max3`.

* Replaced `__swap` to `std::swap`. Removed project macro `__swap`.

* Using `std::extent<decltype(arr)>` for `NELEM(arr)` in C++; removed type casting.

* Applying `override` specifier instead `virtual` for overridden methods.

* Bugfix: virtual destructor of `class CancellationHandler` was missing.

+++ core/indigo-core/reaction/src/query_reaction.cpp
@@ -157,10 +157,10 @@ int QueryReaction::_compare(int& i1, int& i2, void* c)
     int label2 = context.m.getAtomNumber(i2);
     int idx1, idx2;
 
-    for (idx1 = 0; idx1 < (int)NELEM(labels_by_freq); idx1++)
+    for (idx1 = 0; idx1 < NELEM(labels_by_freq); idx1++)
         if (label1 == labels_by_freq[idx1])
             break;
-    for (idx2 = 0; idx2 < (int)NELEM(labels_by_freq); idx2++)
+    for (idx2 = 0; idx2 < NELEM(labels_by_freq); idx2++)
         if (label2 == labels_by_freq[idx2])
             break;
